Change upgrades to use Scriptable Objects, import TMP fonts
Upgrades are now Scriptable Objects, and share a single prefab.
The generalUpgrade script is now used for all upgrades, and uses SO to work as a specific upgrade.
In order to have automatic wrap on upgrade description, I had to use TextMeshPro (TMP) instead of TextMesh. This required importing some TextMeshPro resources, and making TMP versions of the fonts we want to use in the TMP objects.