Bugfix upgrades not getting picked up when pressing E
Due to the collider/trigger on GeneralUpgrades prefab going to sleep, the OnTriggerStay function that lets you pick up an upgrade by pressing E was unreliable, even though the toggling of the tooltip worked as intended.

The fix: added a rigidbody to the prefab, setting it to Kinematic and Never Sleep.