Add campfire prefab
In prefabs/map there is now a campfire prefab, ready to be placed in the world as is.
It has two particle effects, a wood sprite, and a light that sort of flickers.
The slam-dust particles were Lit until now, leading to some strange results in darkness, so they've been changed to an unlit shader now. Thanks to this, they will look the same no matter the surrounding light.