Adjust lights, add crow disappear effect, change to sontinuous collision
Adjusted the lights, making player light smaller and darker, but sharper and stronger. The global light is a little brighter.
Changed alfs and trolls to continuous collision detection after playtesting in builds (they were walking through walls too easily).
Made a particle effect that triggers when the crows from crowbolt hit something or disappear, for better visual feedback.
The "press space to play" bounce was frame dependant and super slow in builds, so it's now been moved to FixedUpdate.